+++ src/oxl_ansible_webui/aw/api.py
@@ -9,7 +9,7 @@ from aw.api_endpoints.credentials import APIJobCredentials, APIJobSharedCredenti
     APIJobSharedCredentialsItem, APIJobUserCredentialsItem, APIJobUserCredentials, APIJobTMPCredentials
 from aw.api_endpoints.filesystem import APIFsBrowse, APIFsExists
 from aw.api_endpoints.system import APISystemConfig, APISystemEnvironment, APIUserPasswordChange, \
-    APISSHHostkey, APISSHHostkeyItem, APISSHHostkeyFile, APISSHHostkeyFileItem
+    APISSHHostkey, APISSHHostkeyItem, APISSHHostkeyFile, APISSHHostkeyFileItem, APIUserLanguage
 from aw.api_endpoints.repository import APIRepository, APIRepositoryItem, APIRepositoryLogFile
 from aw.api_endpoints.alert import APIAlertPlugin, APIAlertPluginItem, APIAlertUser, APIAlertUserItem, \
     APIAlertGlobal, APIAlertGlobalItem, APIAlertGroup, APIAlertGroupItem, APIAlertAll
@@ -59,6 +59,7 @@ urlpatterns_api = [
     path('api/ssh-hostkey-file', APISSHHostkeyFile.as_view()),
     path('api/ssh-hostkey-file/<str:name>', APISSHHostkeyFileItem.as_view()),
     path('api/user/password', APIUserPasswordChange.as_view()),
+    path('api/user/language', APIUserLanguage.as_view()),
     path('api/fs/browse/<int:repo_id>', APIFsBrowse.as_view()),
     path('api/fs/exists', APIFsExists.as_view()),
     path('api/inventory/list', APIInventoryList.as_view()),