+++ src/oxl_ansible_webui/aw/config/language.py
@@ -8,3 +8,9 @@ TRANSLATIONS = {
     'de': DE,
     'es': ES,
 }
+
+LANGUAGE_CHOICES = {
+    'en': 'English',
+    'de': 'Deutsch',
+    'es': 'Español',
+}